backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
pydantic-settings==2.9.1
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
